Split up plugin code into separate module
